fix: Fix wrong types

+++ array.h
@@ -139,7 +139,7 @@ namespace original{
     {
         if (this == &other) return true;
         if (this->size() != other.size()) return false;
-        for (int i = 0; i < this->size(); i ++)
+        for (size_t i = 0; i < this->size(); i ++)
         {
             if (this->body[i] != other.body[i])
             {
@@ -197,7 +197,7 @@ namespace original{
     template <typename TYPE>
     auto original::array<TYPE>::indexOf(const TYPE &e) const -> size_t
     {
-        for (int i = 0; i < this->size(); i += 1)
+        for (size_t i = 0; i < this->size(); i += 1)
         {
             if (this->get(i) == e)
             {
